# openapi.yaml
openapi: "3.1.0"
info:
  title: app-catalogo
  description: |
    Discovery layer unificada da Prefeitura do Rio de Janeiro.
    Indexa serviços públicos (Carta de Serviços via SalesForce), cursos, vagas de emprego
    e oportunidades MEI com busca full-text e recomendação personalizada por perfil de cidadão.

    ## Autenticação

    O Istio valida o JWT e injeta o header `X-Auth-Request-Token` (Bearer JWT).
    O claim `preferred_username` no JWT contém o CPF do cidadão (apenas dígitos).

    Rotas marcadas com `bearerAuth` requerem esse header.
    Rotas marcadas com `adminAuth` requerem adicionalmente a role `admin` ou `go:admin`
    no claim `resource_access.superapp.roles`.

    ## Rate Limiting

    Limite global de **300 requisições/minuto por IP**. Excedido, retorna `429`.

    ## Headers de Resposta

    Todas as respostas incluem `X-Request-ID` com um UUID único por requisição.
  version: "1.0.0"
  contact:
    name: Prefeitura do Rio de Janeiro
    url: https://github.com/prefeitura-rio/app-catalogo
servers:
  - url: https://api.example.com
    description: Produção
  - url: http://localhost:8080
    description: Local
...
security: []
paths:
...
  /api/v1/recommendations:
    get:
      tags: [Recomendações]
      summary: Recomendações personalizadas
      description: |
        Recomendações baseadas no perfil do cidadão autenticado (escolaridade, renda,
        localização, acessibilidade, faixa etária). O perfil é obtido do app-rmi
        on-demand com refresh em background quando estale.

        O campo `score_breakdown` detalha a contribuição de cada dimensão no score final.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/types"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/context"
      responses:
        "200":
          description: Recomendações personalizadas
          headers:
            X-Request-ID:
              $ref: "#/components/headers/X-Request-ID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecommendationResponse"
              examples:
                personalizado:
                  summary: Resposta com perfil encontrado
                  value:
                    personalized: true
                    context: "homepage"
                    items:
                      - id: "550e8400-e29b-41d4-a716-446655440000"
                        type: "service"
                        source: "salesforce"
                        title: "Bolsa Família"
                        short_desc: "Transferência de renda para famílias em vulnerabilidade"
                        organization: "Secretaria de Assistência Social"
                        url: "https://prefeitura.rio/bolsa-familia"
                        modalidade: "presencial"
                        bairros: ["Centro", "Tijuca"]
                        tags: ["assistência", "renda"]
                        score: 0.82
                        score_breakdown:
                          escolaridade: 0.17
                          renda: 0.20
                          localizacao: 0.20
                          acessibilidade: 0.15
                          faixa_etaria: 0.08
                          tipo: 0.02
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/public/recommendations:
    get:
      tags: [Recomendações]
      summary: Recomendações anônimas
      description: |
        Recomendações sem autenticação. Usa scoring com valores neutros.
        O parâmetro `cluster_hint` aceita bairro ou faixa etária para
        leve personalização sem identificação do cidadão.

        `personalized` é sempre `false` neste endpoint.
      security: []
      parameters:
        - $ref: "#/components/parameters/types"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/context"
        - name: cluster_hint
          in: query
          required: false
          schema:
            type: string
            examples:
              - "Tijuca"
              - "25-34"
          description: Bairro ou faixa etária para personalização anônima
      responses:
        "200":
          description: Recomendações anônimas
          headers:
            X-Request-ID:
              $ref: "#/components/headers/X-Request-ID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecommendationResponse"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"

  # ---------------------------------------------------------------------------
  # CATÁLOGO
  # ---------------------------------------------------------------------------
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT injetado pelo Istio via header `X-Auth-Request-Token`.
        O claim `preferred_username` deve conter o CPF do cidadão (apenas dígitos).
    adminAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Mesmo JWT do `bearerAuth`, mas o claim `resource_access.superapp.roles`
        deve conter `"admin"` ou `"go:admin"`.

  # ---------------------------------------------------------------------------
  # Parameters
  # ---------------------------------------------------------------------------
  parameters:
...
    types:
      name: types
      in: query
      required: false
      schema:
        type: array
        items:
          $ref: "#/components/schemas/ItemType"
      style: form
      explode: true
      description: Filtro por tipo de item. Múltiplos valores aceitos.
    limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 50
        default: 10
      description: Número de recomendações. Máximo 50.
    context:
      name: context
      in: query
      required: false
      schema:
        $ref: "#/components/schemas/RecommendationContext"
      description: Contexto de uso das recomendações

  # ---------------------------------------------------------------------------
  # Headers
  # ---------------------------------------------------------------------------
  headers:
    X-Request-ID:
      schema:
        type: string
        format: uuid
      description: UUID único da requisição para rastreabilidade

  # ---------------------------------------------------------------------------
  # Responses reutilizáveis
  # ---------------------------------------------------------------------------
  responses:
...
    Unauthorized:
      description: Token ausente ou inválido
      headers:
        X-Request-ID:
          $ref: "#/components/headers/X-Request-ID"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "autenticação necessária"
...
    TooManyRequests:
      description: Limite de 300 req/min por IP excedido
      headers:
        X-Request-ID:
          $ref: "#/components/headers/X-Request-ID"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalError:
      description: Erro interno do servidor
      headers:
        X-Request-ID:
          $ref: "#/components/headers/X-Request-ID"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    ItemType:
      type: string
      enum:
        - service
        - course
        - job
        - mei_opportunity
      description: |
        - `service`: Serviço público da Carta de Serviços (SalesForce)
        - `course`: Curso de capacitação (app-go-api)
        - `job`: Vaga de emprego (app-go-api)
        - `mei_opportunity`: Oportunidade para MEI (app-go-api)
    ItemSource:
      type: string
      enum:
        - salesforce
        - courses
        - jobs
        - mei
        - app-go-api
      description: |
        Fonte de origem do item:
        - `salesforce`: Carta de Serviços
        - `courses`: Cursos do app-go-api
        - `jobs`: Vagas do app-go-api
        - `mei`: MEI do app-go-api
        - `app-go-api`: fonte genérica do app-go-api (não atribuída a tipo específico)
...
    RecommendationContext:
      type: string
      enum:
        - homepage
        - after_search
        - profile
      default: homepage
      description: |
        Contexto de uso que influencia o scoring:
        - `homepage`: página inicial
        - `after_search`: após uma busca
        - `profile`: página de perfil do cidadão
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Mensagem de erro em português

    # --- Busca ---
...
    RecommendationResponse:
      type: object
      required: [items, context, personalized]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/RankedItem"
        context:
          $ref: "#/components/schemas/RecommendationContext"
        personalized:
          type: boolean
          description: |
            `true` quando o perfil do cidadão foi encontrado e usado no scoring.
            `false` quando o perfil não existe ou no endpoint público.
    RankedItem:
      type: object
      required: [id, type, source, title, score]
      properties:
        id:
          type: string
          format: uuid
        type:
          $ref: "#/components/schemas/ItemType"
        source:
          $ref: "#/components/schemas/ItemSource"
        title:
          type: string
        short_desc:
          type: string
        organization:
          type: string
        url:
          type: string
          format: uri
        image_url:
          type: string
          format: uri
        modalidade:
          type: string
        bairros:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        score:
          type: number
          format: float
          minimum: 0
          maximum: 1
          description: Score total de relevância/personalização, normalizado entre 0 e 1.
        score_breakdown:
          type: object
          description: |
            Contribuição de cada dimensão ao score total.
            Útil para debug e explicabilidade da recomendação.
          properties:
            escolaridade:
              type: number
              format: float
              description: "Peso: 0,25"
            renda:
              type: number
              format: float
              description: "Peso: 0,20"
            localizacao:
              type: number
              format: float
              description: "Peso: 0,20"
            acessibilidade:
              type: number
              format: float
              description: "Peso: 0,15 — considera PCD e necessidades especiais"
            faixa_etaria:
              type: number
              format: float
              description: "Peso: 0,10"
            tipo:
              type: number
              format: float
              description: "Peso: 0,10 — afinidade entre tipo do item e contexto da requisição"

    # --- Catálogo ---
...